requirements.txt

flask
requests
cloudinary
pillow
pillow-heif
